Version 0.75
Completely rewrote the Animation object. It is now a child of the Timer
object and has a few loop types added. Also cleaned some documentation,
added more documentation to the Timer object, and changed some stuff so
it's children can interface with it with more ease. Also added a
background to the debug text, added global timers to the default debug
text, in addition to making the debugging red box not show up for the
debugText.