Version 0.76
Added a new event system.
Changed the events in the animation and timer objects to use the new
system.
Changed all objects that were hooked to timer events to use the new
system.
Reformatted ObjectBase documentation to the new documentation format.
Added a "Size" field to ObjectBase.
Cleaned up the AndroidBuild.