Version 0.94a
Non-defined texture objects will now correctly assign the image name, as
well as the missing texture image.

Improved TextObject justification further to not apply to the last line
of text, and not be centered under JustifiedCenter.

Added a check to the TextObject if the text image is ready, to prevent
flickering.

Added CenterX and CenterY functions to the ObjectBase.