Version 0.94b
Changed the names of the Core update and draw events, and added a new
event that is executed when the window is closed.

The engine will no longer run its update cycle if not focused,
preventing input from other programs from being captured.

Fixed a bug with the TextObject when it has no text and is set to
justify.