Version 94, Patch 1
Fixed the background of the debug and fps texts not displaying.

Made the parent property public.

Reverted the UIObjectBase adding its update function to the update event
as that causes it to trigger events even when the screen it is on is not
loaded, this will be fixed in an upcoming update that adds an object
manager to screens.