Version 94e
Moved the order of stuff in the Core update function to make sure debug
text can be changed by user code.

Added some error prevention to the map object.

Changed private members of the ObjectBase to protected.

Added a smooth move function to the objectbase, and removed the Move
effect object.

Changed some private properties of the physicsobject not to conflict
with existing properties of objectbase.

The textobject will no longer leak memory when changed often.

Added an option to the timer not to start immediately.