Version 3 Development
Including a prebuilt version of the Consolas font for systems that might
not have the font installed.
Increased the accuracy of the ActiveText's justification and
calculations.
Cleaned up ActiveText cache code.
Text will now be cached instantly when the property is changed, which in
turn means the scroll bottom functions doesn't have to be queued.