Version 5 Development
Added a StartRenderTarget overload that will skip redefining and size
checks.
Added a script function for selecting an object.
Changed all inheritance of component functions to be optional.
Fixes and cleanup on the ActiveTexture component., and exposed internal
texture.
Remade the Animation component.
Separated object bounds debugging from texture bounds debugging.
Added an initialization function to the component for initialization
code that needs access to the attachedObject.
Improved removing clusters and objects from the Scene and well as making
sure objects are disposed.
Remade the MouseInput component to be more accurate.
Added the first UI object - the scrollbar.
Improved object listing in the console.