Version 5 Development - Commit text and font typo
Fixed asset integrity checks.
Updated Asset-Meta-Generator to version 4.
Added fade in and fade out effects to the sound engine.
Fixed the debugging console interfering with UI objects.
The animation component will no longer reset if being set a new starting
or ending frame which is the same.
Fixed the TiledMap component not drawing through DrawComponent.
Changing the color of an ActiveText component will now cause a recache.
Clusters will now be taken in consideration when calculating UI object
overlapping, but only if the scene's UIClusters bool is set to true.
Added a tags argument to tickers.
The button UI object will now correctly pass the MouseButtonEventArgs on
the OnClicked event.
The MouseInput component will no longer perform events if its parent is
hidden.
The scrollbar UI object should now be more accurate and now has a
OnValueChanged event, as well as not render its children when hidden.