Version 6 Development
Added Tile and Frame modes to the ActiveTexture component.
Remade the Animation component to only deal with Animation. Framesize,
spacing and cutting is now based in the ActiveTexture component using
the new Tile and Frame modes.
Added component priority back.
ActiveText components will now generate cache upon Initialization.
Added a on click event to the Scene.
Added the ability for box objects to foster children objects under the
box component.
Pulled out the inObject and CheckOpacity functions of the MouseInput
component to the Functions library.