Version 6 Development
Assets will now unload the ContentManager.
Moved all module setting checks to inside to module itself to keep Core
initialization clean.
Fixed a bug where the ActiveTexture component when set to animation
would display the whole spritesheet for a couple of frames.
Fixed a bug where the ActiveText component would not regenerate it's
cache when the object's size changes.
Added a setting for turning scripting on and off.