Version 6 Development
Fixed a bug where the debug console would scroll in the wrong direction.
Fixed a bug where the animation component would not start from the
starting frame.
Fixed a bug where the box component would not render correctly.
FIxed a bug where components would not render correctly when scaled or
when a drawing area is selected.