UI and Input fixes.
UI Control's MouseEnter and MouseLeave events will now consume the position of the mouse.
RichText's Reset function can now be overriden.
Added a new function to the TextDrawingSession to return the bounds of the next glyph.
Input will no longer capture the first mouse click when focused.
Added a Focused boolean to the SDLWindow class.