Sound engine and various fixes.
Added a sound engine based on OpenAL using the OpenTK bindings.
 - Added OpenAL for Windows x86 and x64
 - Added FadeIn and FadeOut effects.
 - Added loading of .Wav files as a new SoundFile asset type.
 - Added SoundManager to the Debugger sources.
 - Added the SoundManager class to the Context.
 - Added example for playing sounds and effects.
While the window is not focused the engine will not draw anything. In addition layer's update cycles won't be run, but a new LightUpdate function will be in its place.
Added a small interrupt to prevent high CPU usage in the Update and Draw cycles. Does not damage performance.
Added documentation to the Layer class.
Fixed layer Unload not being called.
Fixed the UI Controller invoking clicks on controls when the mouse is held on another control and then moved.
Fixed mouse input triggering on focus gain.
Updates to the ReadMe.