Expanded debugging and fixed some bugs.
The debugger now has a type and source filter.
 - Changed some trace type logs to info.
 - At start all sources will be logged, and all types except traces.
Added debugging function for dumping information about a UI controller.
In debug mode all UI controllers will have their controls renderered. Red if inactive and green if active.
 - Added debugging hook in the Controller class.
Added UIController as a debug message source and added debugging traces and infos in the class.
The RichText's update method will now return a boolean signifying whether the scrolling effect has updated or not.
Renamed the UI.Control's Remove() method to Destroy() as that is what it does.
Added a Remove method to the UI.Controller, for removing and destroying a control's reference. Destroyed controls will still be automatically removed, but a warning will be displayed.
The SoundManager's PlaySoundLayer method will no longer accept an argument to loop the source as that is a property of the source which can be modified by the user at a later time, instead it will accept a starting source volume as that is required by the constructor. Default is 1f.
Renamed Source properties to follow convention. IsAdjective -> Adjective.
Added a FileName property to the Source to allow checking which file the source is playing.
Added a helper for generating a random integer using a global seed.
Updated SoundExample to reflect changes.