Fixed Mac build, debugger and shader improvements.
Changed the AppleRun script to address an issue with Freetype and libpng.
Added seconds and milliseconds to the debugger file logs.
Made sure GLProgram uniforms are uploaded on the GL thread.
The renderer will now provide a "time" uniform to the shader. This is the total time passed since the start and can be used for various effects.