Sound fixes.
When setting a new sound source for a layer it will play immediately instead of at the next update. In addition the old source will be destroyed after the old one has been updated.
Destroying a source of a layer which has never played anything will no longer cause a crash.
Added a trace for when a source has finished playing.