Bug fixes.
Changed Input example.
Added the ability to multiply colors by a float which will multiply the alpha.
Updated the Font's MeasureString function to take lines into consideration and to be more accurate.
Added functions for getting key states in the Input module from key codes.
Added a function for converting a key code to a key name.
Fixed a bug with the streaming source where it would play after destroyed and cause a crash.
Source sources will now once again be destroyed before a new one is set and not after.
Added a redudancy check to the TextDrawingSession.