Bug fixes, sound improvements.
When getting a sound layer if it doesn't exist it will be automatically created.
The ALThread will now not block on execution.
The AssetLoader will now reject empty paths.
Added the ability to get the stripped text from the RichText class.
You can now highlight debug messages from a specified source.
Added the "debugAudio" command which dumps the status of all sound layers.