SoundManager improvements and some Async Queuing added to it
Fixed FadeIn not playing when FadeInFirstLoopOnly was turned on and you played a sound without appending it.
Added the ContinuousAction class which allows a function to return a ContinuousAction object which another action can be appended to. This is a temporary simplistic solution and should be replaced with C# Tasks down the line.
Changed the SoundManager's functions to return ContinuousAction objects.