Various improvements related to the Host, and bug fixed with UI and sound.
Code Style applied.
Moved Host namespace to Emotion.Engine.Hosting
Renamed Window to OtkWindow (OpenTK Window) and added a new Desktop namespace for it.
 - Removed the RenderSize property. Use the one in Context.Settings instead.
 - Moved the Resize code to the Context.
 - The Context will no longer assume the Host's Run function is blocking.
Scripting errors will now be logged as such instead of as traces and warnings.
Added a RenderSize Vector2 to the settings object, this is basically RenderWidth and RenderHeight combined. The two components are also now floats instead of ints.
The UI controller will no longer block mouse clicks if the mouse down event of a control deactivates it.
Moved the OpenGLMajor and Minor version flags to the Renderer
Added comments and documentation to some places.
The SoundLayer's FadeIn and FadeOut lengths are now floats.
Fading will now work on MacOS.
 - The getting of the current buffer playing on the source is not working on Mac for some reason.
ForceFadeOut on loops will now catch up if missed.
OpenAL errors will still be ignored on MacOS but will now be logged. This might lead to log spam.