Improved scaling.
All rendering is now done to a FBO of the RenderSettings size which is then rendered to the main FBO.
 - This should prevent weird pixel scaling on non-integer scales.
Added an overload to the texture creator to create a texture from an OGL pointer.
Removed forgotten debug code in the TileMap.
Borderless will be set to Fullscreen on Mac as well as Linux because of buggy behavior.