Improvements and Fixes
Automatic scaling will now affect the host size as well as the render size.
Textures are now loaded upside down and then rotated using an inverted model matrix. This is because external plugins and renderers expect the textures to be upside down, and only the Adfectus renderer knows this to not be the case.