AStar and Camera Fixes
The view matrix is now managed by the Renderer rather than the GraphicsManager.
The AStar grid, node, and path are now AStarGrid, AStarNode, and AStarPath.
You can now store data in the AStarNode instance.
The AStarPath will now no longer go out of bounds when searching for neighbours.
Moved the ScreenToWorld function from the Renderer to the Camera, and added a WorldToScreen function.