Various Features + RenderTargets, Removed support for OGL 2
OpenGL 2 is no longer semi-supported.
   - The shaders remain for now.
   - Removed flags include: TextureLoadStandard, UseVao, UseFramebuffer, and SetVertexAttribLocations. These are all on by default now.
Added StateClip(bool) to the GraphicsManager which allows you to trigger ScissorTesting off and on.
Added support for RenderTargets!
  - The scale FBO is now the default RenderTarget.
  - Render targets are implemented as a stack similar to the model matrix.
The ViewMatrix now uses the "center" overloads of the scale matrix instead of calculating it itself.
The default QuadIbo is now an accessible property.
Added a default ibo for drawing vertices.
Added a ModelMatrix and CurrentTarget properties to the renderer to let you check what is on top of the stacks currently.
Renderer.RenderVertices now actually renders vertices instead of quad-like vertex arrays.
Added a new Rectangle constructor.
The SceneManager will no longer lock on the swapMutex on every call, now it will do it only on the Update loops.
Added a function to MathExtension for checking the BoundingRectangleOfPolygon.