TileMap Tuneup and RenderSize Deprecation
RenderSize has been deprecated in favor of Engine.Renderer.CurrentTarget.Size
The Adfectus.Game.Tiled.Map class has been renamed to TileMap, and its constructors have been revised to better fit the engine.
Added a BaseTarget property to the Renderer which points to the render target on the bottom of the stack.