ImGui Fixes, Bug Pixels, Tooling
Applied code style to Plugins.ImGuiNet.
Fixed some problems with scaling and ImGui not rendering window contents.
Added an extension for getting a texture's UVs for use in ImGui.Image.
Added Plugins.ImGuiNet.Windowing as a namespace and the ImGuiWindow and WindowManager classes which provide a more OOP way of managing ImGui windows.
Added Emotion.Tools. This project will contain various development tools and will aim to be both a toolset and a plugin to Emotion.
  - Added an animation editor/previewer to present how Emotion.Tools can be used as a toolkit.
  - Added a performance profiler to present how Emotion.Tools can be used as a plugin. It still isn't importable as a plugin though.
Fixed a race condition with the _unfocusedHostEvent.
Expanded AnimatedTexture
 - You can now request the frame bounds of any frame index.
 - New constructor allows you to specify rows and columns instead of a frame size.
 - Frame size can now be changed after initialization.
 - You can now force set a specific frame.
The AssetLoader now exposes an API for getting the source which contains a specific asset. This is intended to be used for debugging and such.
Updated to the ReadMe about features and used projects.