Optimization of textures within the engine.
Emotion.Standard.Image will no longer output, and accept images as RGBA with a top origin, but rather BGRA with a bottom origin - the way it is stored inside the image itself. It is the concern of the user to convert that.
  - Emotion.Test now handles this
  - TextureAsset now handles this
ImageUtil.FlipImageY now only accepts the parameters it actually uses, and added a variation of the function which doesn't mutate the input pixel data.
Fixed drawing RenderSpriteCommand without a batch and without a texture.
Images internally in the engine are now BGRA