Fixes and Features
The "Scale" property is now equal to the smaller ratio between X and Y, rather than being the ratio between the smaller target dimension.
The XMLAsset now has a function to serialize an object.
The SceneManager will now throw any exceptions in the Load and Unload functions if a debugger is attached.