Generic Batch and Shared Memory Batch
Added a new NativeMemoryPool class.
The RenderComposer now owns native memory which can be used by other commands. If none if it is used, it isn't allocated.
RenderComposer will now dispose of cached vaos.
The SpriteBatch is now generic and can be used for structs other than VertexData.
Added a new SharedMemorySpriteBatch which uses the new memory in the render composer - meaning less memory usage on state changes. It is also the default batch type now.