Configuration Refactor, Custom Loops, Variable Loop Speed
Refactored the way the configuration is setup from a "AspNet like builder" to an object with properties.
Added the ability to specify a custom game loop.
The default game loop will now run twice as fast or up to four times as slow if it cannot keep up. This is configurable.
Moved VSync option to the Renderer and ForcedVSync to it.
Added Reported DeltaTime to the PerformanceMonitor.
Emotion.Test now uses a custom loop.
FullScale is now the default scaling mode. The letter/pillarbox mode is called ScaleBlackBars.
Moved RenderDocAPI to its own file.
Added a property to the SceneManager for whether it is currently on the loading screen.