[Text] Larger fonts will be snapped to the pixel grid.
This is kind of a regression since gPos was always rounded before, but since the SDF commit I'm trying to make text more "accurate". Of course that means that it doesn't snap neatly into the pixel grid, causing some glyphs to be wider than others, which looks abysmal. Snapping is applied only to the X, since rounding the Y will cause the baseline to jump.

Other changes is code style being applied.