[Rendering] FencedBufferSource will now use MapBufferRange on Intel hardware
Something which looks like a driver bug occurs on Intel Xe when using BufferSubData. A single flickering frame is displayed randomly. It doesn't occur if you allocate a buffer every time you flush. Maybe the driver doesn't copy the contents of the buffer to pinned memory in some very specific cases and uses it as is and when it gets written to junk is uploaded? Fences should prevent this though...